Update README.md
In this project, a machine learning model will be developed to predict loan eligibility based on applicant details such as income, credit history, and other financial factors. The dataset will be explored to understand its structure and key variables. Various machine learning algorithms will be implemented and compared to determine the best-performing model. This project offers practical experience in predictive modeling and working with real-world datasets.